chore(tip): merges retip & tip logics
