+++ modules/comment/logic/commentLogic.js
@@ -1,7 +1,6 @@
-const { Comment, Profile } = require('../../../models');
+const { Comment, Profile, Tip } = require('../../../models');
 const NotificationLogic = require('../../notification/logic/notificationLogic');
 const cache = require('../../cache/utils/cache');
-const TipLogic = require('../../tip/logic/tipLogic');
 const { NOTIFICATION_TYPES } = require('../../notification/constants/notification');
 
 module.exports = class CommentLogic {
@@ -19,7 +18,7 @@ module.exports = class CommentLogic {
         return res.status(400).send(`Could not find parent comment with id ${parentId}`);
       }
 
-      const relevantTip = await TipLogic.getOne(tipId);
+      const relevantTip = await Tip.findOne(tipId);
       if (!relevantTip) return res.status(400).send(`Could not find tip with id ${tipId}`);
 
       const entry = await Comment.create({
